requirements.txt
# see ./requirements-large.txt

waivek
wheel
pytest
Flask
flask-cors
pysqlite3-binary
timeago
python-dotenv
Flask-SQLAlchemy
SQLAlchemy-serializer
python-slugify

# chat.py
requests
requests_ip_rotator

# download_worker
psutil
yt-dlp

# test
pydantic

# video_stats
pandas

# videos.py
isodate
diskcache

# pov_isolation.py
opencv-python

# for url to domain conversion for get_gateway
furl
